rename read-file to keychain-read-file
instead of only defining it if not already define.  If read-file is part
of Emacs then it will likely be different than yours anyway and your
library will break if you use it, let's be prepared for that and also
be nice and use library prefix like you ought to.

Also cleanup function a bit while we are at it.