add variable keychain-directory
Remove the variables keychain-ssh-file and keychain-gpg-file replacing
them with variable keychain-directory.  keychain actually allows
changing the directory but not the filenames of files stored within.

Also I cannot find a way to only create files for the shell actually
being used (one of sh, csh or fish).  So we can continue to assume
that the sh-variants exist and don't have to let the user select
another.