keychain-environment: call keychain directly
Pull request #2 broke keychain on my Ubuntu LTS system. It seems there
is some debate over what value of ${HOSTNAME} is used when creating the
file paths. To avoid this ambiguity we can call keychain directly with
the --eval option. As a bonus we can delete the rest of the code for
reading the files into buffers.

Signed-off-by: Alex Bennée <alex.bennee@linaro.org>